Move and add test
